docs: v1.0 ship-blocker audit, platform matrix, and .stune format spec (#1)
v1.0 ship-blocker audit, platform feature matrix, .stune directory-format spec,
brick-protection per-ISA split (SH-2A vs RH850), cancellation-invariant
formalization (mid-PDU deferral + crash-recovery), [[table.role]] schema
promoted to v1.0, OFL font notices, README platform matrix, plus the CI
matrix hardening that fell out of getting it green (GCC -Wnull-dereference
+ -Wformat-truncation fixes, clang -Wdouble-promotion fixes, missing
serial_byte_channel POSIX stub, and a sweep replacing Unicode arrows /
em dashes in Catch2 test names because Windows CTest mangles them).

+++ src/feature_codegen/src/sh2a.hpp
@@ -49,29 +49,33 @@ enum class Reg : std::uint8_t {
 // — undefined behaviour means the emitter must range-check before
 // calling.
 [[nodiscard]] constexpr std::uint16_t enc_mov_l_disp_pc(Reg rn, std::uint8_t disp) noexcept {
-    return static_cast<std::uint16_t>(0xD000U | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      static_cast<std::uint16_t>(disp));
+    return static_cast<std::uint16_t>(
+        0xD000U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(disp));
 }
 
 // MOV.L Rm, @Rn  — store the 32-bit value in Rm to the address in Rn.
 // Encoding: 0010 nnnn mmmm 0010.
 [[nodiscard]] constexpr std::uint16_t enc_mov_l_reg_at_reg(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x2002U | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x2002U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // MOV.L @Rm, Rn  — load the 32-bit value at the address in Rm into Rn.
 // Encoding: 0110 nnnn mmmm 0010.
 [[nodiscard]] constexpr std::uint16_t enc_mov_l_at_reg_reg(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x6002U | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x6002U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // ADD Rm, Rn  — Rn = Rn + Rm. 32-bit signed (or unsigned — two's-
 // complement wrap is identical). Encoding: 0011 nnnn mmmm 1100.
 [[nodiscard]] constexpr std::uint16_t enc_add(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x300CU | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x300CU | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // SUB Rm, Rn  — Rn = Rn - Rm. Encoding: 0011 nnnn mmmm 1000.
@@ -79,8 +83,9 @@ enum class Reg : std::uint8_t {
 // so the result lands in Rn following the codegen's "result in R1"
 // convention.
 [[nodiscard]] constexpr std::uint16_t enc_sub(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x3008U | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x3008U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // MUL.L Rm, Rn  — MACL = Rn * Rm (low 32 bits of the 64-bit signed
@@ -88,22 +93,25 @@ enum class Reg : std::uint8_t {
 // MACL system register, not a GPR — callers must follow up with
 // STS MACL, Rn to extract the value. Encoding: 0000 nnnn mmmm 0111.
 [[nodiscard]] constexpr std::uint16_t enc_mul_l(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x0007U | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x0007U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // STS MACL, Rn  — Rn = MACL. Used after MUL.L to copy the low 32
 // bits of the multiplication result into a general-purpose register.
 // Encoding: 0000 nnnn 0001 1010.
 [[nodiscard]] constexpr std::uint16_t enc_sts_macl(Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x001AU | (static_cast<std::uint16_t>(rn) << 8U));
+    return static_cast<std::uint16_t>(
+        0x001AU | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U));
 }
 
 // CMP/EQ Rm, Rn  — T = (Rn == Rm). Commutative — operand swap
 // produces the same result. Encoding: 0011 nnnn mmmm 0000.
 [[nodiscard]] constexpr std::uint16_t enc_cmp_eq(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x3000U | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x3000U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // CMP/GT Rm, Rn  — T = (Rn > Rm) signed. NOT commutative; callers
@@ -111,8 +119,9 @@ enum class Reg : std::uint8_t {
 // for "R1 > R0 = b > a = a < b") or `>` (CMP/GT R1, R0 for
 // "R0 > R1 = a > b"). Encoding: 0011 nnnn mmmm 0111.
 [[nodiscard]] constexpr std::uint16_t enc_cmp_gt(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x3007U | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x3007U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // MOVT Rn  — Rn = T (zero-extended; result is 0 or 1). Materializes
@@ -120,31 +129,35 @@ enum class Reg : std::uint8_t {
 // stored to memory like any other SSA value. Encoding:
 // 0000 nnnn 0010 1001.
 [[nodiscard]] constexpr std::uint16_t enc_movt(Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x0029U | (static_cast<std::uint16_t>(rn) << 8U));
+    return static_cast<std::uint16_t>(
+        0x0029U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U));
 }
 
 // AND Rm, Rn  — Rn = Rn & Rm. Bitwise AND of two 32-bit registers.
 // On canonical 0/1 Bool inputs, produces canonical 0/1 logical AND.
 // Encoding: 0010 nnnn mmmm 1001.
 [[nodiscard]] constexpr std::uint16_t enc_and(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x2009U | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x2009U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // OR Rm, Rn  — Rn = Rn | Rm. Bitwise OR. On canonical 0/1 Bool
 // inputs, produces canonical 0/1 logical OR.
 // Encoding: 0010 nnnn mmmm 1011.
 [[nodiscard]] constexpr std::uint16_t enc_or(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x200BU | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x200BU | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // TST Rm, Rn  — T = ((Rn & Rm) == 0). Used as the "is zero" test for
 // logical NOT: TST Rn, Rn followed by MOVT flips 0↔1. Encoding:
 // 0010 nnnn mmmm 1000.
 [[nodiscard]] constexpr std::uint16_t enc_tst(Reg rm, Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x2008U | (static_cast<std::uint16_t>(rn) << 8U) |
-                                      (static_cast<std::uint16_t>(rm) << 4U));
+    return static_cast<std::uint16_t>(
+        0x2008U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 4U));
 }
 
 // BT label  — branch if T=1. NO delay slot (the /S variant has one).
@@ -198,21 +211,24 @@ enum class FReg : std::uint8_t {
 
 // FADD FRm, FRn — FRn = FRn + FRm. Encoding: 1111 nnnn mmmm 0000.
 [[nodiscard]] constexpr std::uint16_t enc_fadd(FReg frm, FReg frn) noexcept {
-    return static_cast<std::uint16_t>(0xF000U | (static_cast<std::uint16_t>(frn) << 8U) |
-                                      (static_cast<std::uint16_t>(frm) << 4U));
+    return static_cast<std::uint16_t>(
+        0xF000U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(frm) << 4U));
 }
 
 // FSUB FRm, FRn — FRn = FRn - FRm. NOT commutative.
 // Encoding: 1111 nnnn mmmm 0001.
 [[nodiscard]] constexpr std::uint16_t enc_fsub(FReg frm, FReg frn) noexcept {
-    return static_cast<std::uint16_t>(0xF001U | (static_cast<std::uint16_t>(frn) << 8U) |
-                                      (static_cast<std::uint16_t>(frm) << 4U));
+    return static_cast<std::uint16_t>(
+        0xF001U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(frm) << 4U));
 }
 
 // FMUL FRm, FRn — FRn = FRn * FRm. Encoding: 1111 nnnn mmmm 0010.
 [[nodiscard]] constexpr std::uint16_t enc_fmul(FReg frm, FReg frn) noexcept {
-    return static_cast<std::uint16_t>(0xF002U | (static_cast<std::uint16_t>(frn) << 8U) |
-                                      (static_cast<std::uint16_t>(frm) << 4U));
+    return static_cast<std::uint16_t>(
+        0xF002U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(frm) << 4U));
 }
 
 // FDIV FRm, FRn — FRn = FRn / FRm. NOT commutative.
@@ -221,22 +237,25 @@ enum class FReg : std::uint8_t {
 // posture and trust the user to not author a zero-divisor (a
 // divide-by-zero guard primitive is a later concern).
 [[nodiscard]] constexpr std::uint16_t enc_fdiv(FReg frm, FReg frn) noexcept {
-    return static_cast<std::uint16_t>(0xF003U | (static_cast<std::uint16_t>(frn) << 8U) |
-                                      (static_cast<std::uint16_t>(frm) << 4U));
+    return static_cast<std::uint16_t>(
+        0xF003U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(frm) << 4U));
 }
 
 // LDS Rm, FPUL — FPUL = Rm (transfer GPR → FPU communications
 // register). Used as the first step of "move int-bit pattern into
 // FPU side." Encoding: 0100 mmmm 0101 1010.
 [[nodiscard]] constexpr std::uint16_t enc_lds_r_fpul(Reg rm) noexcept {
-    return static_cast<std::uint16_t>(0x405AU | (static_cast<std::uint16_t>(rm) << 8U));
+    return static_cast<std::uint16_t>(
+        0x405AU | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rm) << 8U));
 }
 
 // STS FPUL, Rn — Rn = FPUL (transfer FPU communications register →
 // GPR). Used as the first step of "move FPU result back into the
 // int-side store path." Encoding: 0000 nnnn 0101 1010.
 [[nodiscard]] constexpr std::uint16_t enc_sts_fpul_r(Reg rn) noexcept {
-    return static_cast<std::uint16_t>(0x005AU | (static_cast<std::uint16_t>(rn) << 8U));
+    return static_cast<std::uint16_t>(
+        0x005AU | static_cast<std::uint16_t>(static_cast<std::uint16_t>(rn) << 8U));
 }
 
 // FSTS FPUL, FRn — FRn = FPUL. Bit-cast: copies the 32 bits of
@@ -245,14 +264,16 @@ enum class FReg : std::uint8_t {
 // floating-point register so FADD/FSUB/FMUL/FDIV can operate on it.
 // Encoding: 1111 nnnn 0000 1101.
 [[nodiscard]] constexpr std::uint16_t enc_fsts_fpul_freg(FReg frn) noexcept {
-    return static_cast<std::uint16_t>(0xF00DU | (static_cast<std::uint16_t>(frn) << 8U));
+    return static_cast<std::uint16_t>(
+        0xF00DU | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frn) << 8U));
 }
 
 // FLDS FRm, FPUL — FPUL = FRm. Mirror of fsts: bit-cast FRm into
 // FPUL so STS FPUL, Rn can pull the bits out to a GPR for the
 // memory store. Encoding: 1111 mmmm 0001 1101.
 [[nodiscard]] constexpr std::uint16_t enc_flds_freg_fpul(FReg frm) noexcept {
-    return static_cast<std::uint16_t>(0xF01DU | (static_cast<std::uint16_t>(frm) << 8U));
+    return static_cast<std::uint16_t>(
+        0xF01DU | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frm) << 8U));
 }
 
 // FLOAT FPUL, FRn — FRn = (float32)(int32)FPUL. **Conversion**, not
@@ -262,7 +283,8 @@ enum class FReg : std::uint8_t {
 // Precision: int32 values with |x| > 2^24 may lose low-order bits
 // in the conversion. Encoding: 1111 nnnn 0010 1101.
 [[nodiscard]] constexpr std::uint16_t enc_float_fpul_freg(FReg frn) noexcept {
-    return static_cast<std::uint16_t>(0xF02DU | (static_cast<std::uint16_t>(frn) << 8U));
+    return static_cast<std::uint16_t>(
+        0xF02DU | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frn) << 8U));
 }
 
 // FTRC FRm, FPUL — FPUL = (int32)trunc(FRm). **Conversion**, not a
@@ -273,15 +295,17 @@ enum class FReg : std::uint8_t {
 // outside int32 range produce INT_MIN / INT_MAX, no trap.
 // Encoding: 1111 mmmm 0011 1101.
 [[nodiscard]] constexpr std::uint16_t enc_ftrc_freg_fpul(FReg frm) noexcept {
-    return static_cast<std::uint16_t>(0xF03DU | (static_cast<std::uint16_t>(frm) << 8U));
+    return static_cast<std::uint16_t>(
+        0xF03DU | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frm) << 8U));
 }
 
 // FCMP/EQ FRm, FRn — T = (FRn == FRm). IEEE 754 ordered compare: NaN
 // operands produce T=0. Commutative — operand swap is identical.
 // Encoding: 1111 nnnn mmmm 0100.
 [[nodiscard]] constexpr std::uint16_t enc_fcmp_eq(FReg frm, FReg frn) noexcept {
-    return static_cast<std::uint16_t>(0xF004U | (static_cast<std::uint16_t>(frn) << 8U) |
-                                      (static_cast<std::uint16_t>(frm) << 4U));
+    return static_cast<std::uint16_t>(
+        0xF004U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(frm) << 4U));
 }
 
 // FCMP/GT FRm, FRn — T = (FRn > FRm), signed IEEE 754 ordered. NaN
@@ -290,8 +314,9 @@ enum class FReg : std::uint8_t {
 // FR0)` after loading op1→FR0 / op2→FR1 yields T=(op1 > op2)).
 // Encoding: 1111 nnnn mmmm 0101.
 [[nodiscard]] constexpr std::uint16_t enc_fcmp_gt(FReg frm, FReg frn) noexcept {
-    return static_cast<std::uint16_t>(0xF005U | (static_cast<std::uint16_t>(frn) << 8U) |
-                                      (static_cast<std::uint16_t>(frm) << 4U));
+    return static_cast<std::uint16_t>(
+        0xF005U | static_cast<std::uint16_t>(static_cast<std::uint16_t>(frn) << 8U) |
+        static_cast<std::uint16_t>(static_cast<std::uint16_t>(frm) << 4U));
 }
 
 // RTS — return from subroutine. Encoded as 0x000B.
